+++ src/mapleall/maple_me/src/lmbc_memlayout.cpp
@@ -43,17 +43,17 @@ void LMBCMemLayout::LayoutStackFrame(void) {
     if (sym->GetStorageClass() == kScPstatic && sym->LMBCAllocateOffSpecialReg()) {
       uint32 stindex = sym->GetStIndex();
       sym_alloc_table[stindex].mem_segment = seg_GPbased;
-      seg_GPbased->size = maplebe::RoundUp(seg_GPbased->size, sym->GetType()->GetAlign());
+      seg_GPbased->size = static_cast<int32>(maplebe::RoundUp(seg_GPbased->size, sym->GetType()->GetAlign()));
       sym_alloc_table[stindex].offset = seg_GPbased->size;
-      seg_GPbased->size += sym->GetType()->GetSize();
+      seg_GPbased->size += static_cast<int32>(sym->GetType()->GetSize());
     }
     if (sym->GetStorageClass() != kScAuto) {
       continue;
     }
     uint32 stindex = sym->GetStIndex();
     sym_alloc_table[stindex].mem_segment = &seg_FPbased;
-    seg_FPbased.size -= sym->GetType()->GetSize();
-    seg_FPbased.size = maplebe::RoundDown(seg_FPbased.size, sym->GetType()->GetAlign());
+    seg_FPbased.size -= static_cast<int32>(sym->GetType()->GetSize());
+    seg_FPbased.size = static_cast<int32>(maplebe::RoundDown(seg_FPbased.size, sym->GetType()->GetAlign()));
     sym_alloc_table[stindex].offset = seg_FPbased.size;
   }
 }
@@ -81,9 +81,9 @@ GlobalMemLayout::GlobalMemLayout(MIRModule *mod, MapleAllocator *mallocator)
       }
       uint32 stindex = sym->GetStIndex();
       sym_alloc_table[stindex].mem_segment = &seg_GPbased;
-      seg_GPbased.size = maplebe::RoundUp(seg_GPbased.size, sym->GetType()->GetAlign());
+      seg_GPbased.size = static_cast<int32>(maplebe::RoundUp(seg_GPbased.size, sym->GetType()->GetAlign()));
       sym_alloc_table[stindex].offset = seg_GPbased.size;
-      seg_GPbased.size += sym->GetType()->GetSize();
+      seg_GPbased.size += static_cast<int32>(sym->GetType()->GetSize());
     }
   }
   seg_GPbased.size = maplebe::RoundUp(seg_GPbased.size, GetPrimTypeSize(PTY_ptr));
